+++ backend/canisters/community/api/src/updates/end_video_call.rs
@@ -1,0 +1,16 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use types::{ChannelId, MessageIndex};
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub channel_id: ChannelId,
+    pub message_index: MessageIndex,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+    MessageNotFound,
+    AlreadyEnded,
+}