+++ backend/canisters/group/api/src/updates/end_video_call.rs
@@ -1,0 +1,15 @@
+use candid::CandidType;
+use serde::{Deserialize, Serialize};
+use types::MessageIndex;
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub struct Args {
+    pub message_index: MessageIndex,
+}
+
+#[derive(CandidType, Serialize, Deserialize, Debug)]
+pub enum Response {
+    Success,
+    MessageNotFound,
+    AlreadyEnded,
+}