+++ backend/canisters/group/impl/src/lib.rs
@@ -22,11 +22,11 @@ use std::cell::RefCell;
 use std::collections::hash_map::Entry::{Occupied, Vacant};
 use std::collections::{HashMap, HashSet};
 use std::ops::Deref;
-use types::SNS_FEE_SHARE_PERCENT;
 use types::{
     AccessGate, BuildVersion, CanisterId, ChatMetrics, CommunityId, Cryptocurrency, Cycles, Document, Empty, EventIndex,
     FrozenGroupInfo, GroupCanisterGroupChatSummary, GroupMembership, GroupPermissions, GroupSubtype, MessageIndex,
     Milliseconds, Notification, PaymentGate, Rules, TimestampMillis, Timestamped, UserId, MAX_THREADS_IN_SUMMARY,
+    SNS_FEE_SHARE_PERCENT,
 };
 use utils::consts::OPENCHAT_BOT_USER_ID;
 use utils::env::Environment;
@@ -80,6 +80,11 @@ impl RuntimeState {
 
     pub fn is_caller_escrow_canister(&self) -> bool {
         self.env.caller() == self.data.escrow_canister_id
+    }
+
+    pub fn is_caller_video_call_operator(&self) -> bool {
+        let caller = self.env.caller();
+        self.data.video_call_operators.iter().any(|o| *o == caller)
     }
 
     pub fn is_caller_community_being_imported_into(&self) -> bool {
@@ -216,6 +221,7 @@ impl RuntimeState {
             gate: chat.gate.value.clone(),
             rules_accepted: membership.rules_accepted,
             membership: Some(membership),
+            video_call_in_progress: chat.events.video_call_in_progress.value.clone(),
         }
     }
 
@@ -430,6 +436,13 @@ struct Data {
     pub rng_seed: [u8; 32],
     pub pending_payments_queue: PendingPaymentsQueue,
     pub total_payment_receipts: PaymentReceipts,
+    // TODO: Remove serde default
+    #[serde(default = "video_call_operators")]
+    video_call_operators: Vec<Principal>,
+}
+
+fn video_call_operators() -> Vec<Principal> {
+    vec![Principal::from_text("nmufs-fiu7o-cyg5v-ozcjx-b5qsb-y6nsy-viid6-esfxk-s4nzb-yv2u3-jae").unwrap()]
 }
 
 fn init_instruction_counts_log() -> InstructionCountsLog {
@@ -461,6 +474,7 @@ impl Data {
         test_mode: bool,
         permissions: Option<GroupPermissions>,
         gate: Option<AccessGate>,
+        video_call_operators: Vec<Principal>,
     ) -> Data {
         let chat = GroupChatCore::new(
             creator_user_id,
@@ -503,6 +517,7 @@ impl Data {
             rng_seed: [0; 32],
             pending_payments_queue: PendingPaymentsQueue::default(),
             total_payment_receipts: PaymentReceipts::default(),
+            video_call_operators,
         }
     }