+++ backend/canisters/proposals_bot/impl/src/jobs/check_for_nervous_system_updates.rs
@@ -74,6 +74,10 @@ async fn create_group(ns: NervousSystemDetails, group_index_canister_id: Caniste
                 default: GroupPermissionRole::Admins,
                 ..Default::default()
             },
+            thread_permissions: Some(MessagePermissions {
+                default: GroupPermissionRole::Members,
+                ..Default::default()
+            }),
             ..Default::default()
         }),
         events_ttl: None,